Remove "refactoring" part.
I thought refactoring has a lot in common with autocompletion. It turns
out, that it's not really true. Also there is a UNIX philosophy that
tells us to make tools which do one job and do it well. I wasn't
planning to finish the refactoring stuff anyway, so.. the code is
removed. If you really need it - dig the git history and resurrect the
project in a fork.