Fix nasty bug with transformation of '?' symbols in .a files.
In Go:
	func abc(int)

is the same as:
	func abc(_ int)

On the other hand:
	type ABC struct { _ MyStruct }

is not the same as:
	type ABC struct { MyStruct }

So, I'm kinda trying to apply a hack for that. Therefore I'm changing
'?' to '__' and process it as it's a blank space in case of structs. And
as '_' in case of function arguments.

I need to improve this later and remove that ugly hack.