Fixed test0025: No autocomplete for "unsafe"
Usually, when gocode's parser finds an import line 'import "X"', it
finds out the real file X.a where package X is implemented in, by
issuing the findGlobalFile command, and then, tries to fetch the
information from this file from the package cache.

However, the "unsafe" package is special. It is a builtin package asf
in go, there's no unsafe.a file that gocode can parse. Thus, gocode
generates a synthetic "unsafe" package, and loads it to the package
cache as a package whose name is the result of findGlobalFile("unsafe").
The result of findGlobalFile is a nonexisting file, (usually
~/go/pkg/linux_386/unsafe.a), however it doesn't matter. When the parser
finds 'import "unsafe"', it would run 'findGlobalFile("unsafe")',
receive a nonexisting package name for "unsafe" and would get the
correct synthetic package from the package cache.

I told the findGlobalFile function to check if the package exist, and
notify the user if it doesn't, thus the unsafe package was ignored since
it indeed does not exist. I added therefor an exception for the unsafe
package at the beginning of "findGlobalFile".