Reimplement packages parser (package.go).
Previously, I was relying on a lot of preprocessing hacks to prepare
contents of the .[a568] files for "go/parser". But lately I promised to
reimplement that crappy portion of the gocode in IRC when I have >= 200
followers on the github. Well, here we go.

The new parser is the fork of "go/types".GcImporter. I can't use it
directly, because it has its own representation for declarations and
types. And I in turn, have my own.

New parser parses all the standard library (test.0032, imports all
packages) and performance results are quite nice. On a simple test
with cold cache gocode does autocompletion three times faster
(200ms vs 600ms). Warm cache autocompletion time stays the same
(25-30ms).

"Live long and prosper" (c)