Fix missing variables in type embedding case.
A misinterpreted case, when a type only embeds something and doesn't contain
usual named fields. See example:

type A int
type B { A }

var b B
b.<autocompletion>

Should give A variable, but it was missing in case if there were no ordinary
fields in B. This patch fixes it.