Insert semicolon at the cursor location. Fixes #254.
Often when you do autocompletion, the text your cursor points at is a partial
statement. But go's parser fails to recognize that if it ends with '.' and
continues parsing assuming the statement spans multiple lines (doesn't insert
a semicolon).

This ugly hack fixes it. However, it's unclear how this hack will behave for
cases when you're in the middle of an expression.