Use "AllErrors" flag when parsing Go.
If this flag is not present, "go/parser" will stop parsing after 10 errors.
The standard library Go parser is rather good at returning partially parsed
data and the lack of this flag makes it return nothing. Which, in our case,
leads to lack of autocompletion suggestions.

This case was hard to identify as it happens only when editing rather long
functions and doesn't show up on artificial tests.

Also this patch makes error logging more detailed by using scanner.ErrorList.