Experimental Go 1.5 vendoring support. It's always enabled.
When package-lookup-mode is "go" (which is the default), gocode will try to
figure out a package path you're currently editing. Let's call this package
path - "X". And when importing package using path "Y", it will pretend the
package path you're importing is actually "X/vendor/Y" and only if it fails it
will look for "Y".

Seems like this is what vendoring is all about essentially.